Create a class level bollean variable in the MazeSolver class for flagging when the exit of the maze had been found. This will allow for a much cleaner return to the bottom of the stack call.
Implement a new approach for traversing the maze where an if statement for each direction "looks a head" for a valid pathway, before calling the mazeTraversal method.

Create a method in the MazeSolver class for displaying the programs progression through the maze, coordinate space by coordinate space.

Make prompts for progression through the maze to give the user a more comprehensive interactive experience.